rename ll::common to ll::error
