+++ src/hl/error.rs
@@ -3,9 +3,9 @@
 use std::{borrow::Cow, fmt};
 
 /// A raw [`ll`][crate::ll] error, with metadata such as stack traces.
-pub type LanguageError = crate::ll::common::Error;
+pub type LanguageError = crate::ll::error::Error;
 /// A raw [`ll`][crate::ll] error kind.
-pub type LanguageErrorKind = crate::ll::common::ErrorKind;
+pub type LanguageErrorKind = crate::ll::error::ErrorKind;
 
 /// An error.
 #[derive(Debug)]
