+++ src/ll/codegen/literals.rs
@@ -4,7 +4,7 @@ use super::{CodeGenerator, Expression, ExpressionResult};
 use crate::ll::{
     ast::{Ast, NodeId, NodeKind},
     bytecode::{Opcode, Opr24},
-    common::{Error, ErrorKind},
+    error::{Error, ErrorKind},
 };
 
 impl<'e> CodeGenerator<'e> {
